Revert "reunite svc and basesvc, lay groundwork for auth work (#167)"
This reverts commit 42010cce4b7ba3a86de35c9261fdf9c209310712.
