+++ internal/backend/basesvc/banner.go
@@ -1,4 +1,4 @@
-package svc
+package basesvc
 
 import (
 	_ "embed"
