+++ internal/backend/basesvc/config.go
@@ -1,4 +1,4 @@
-package svc
+package basesvc
 
 import (
 	"errors"
