+++ internal/backend/basesvc/fxsvcs.go
@@ -1,4 +1,4 @@
-package svc
+package basesvc
 
 import (
 	"go.uber.org/fx"
@@ -16,13 +16,13 @@ type fxServices struct {
 	Envs_         sdkservices.Envs         `optional:"true"`
 	Events_       sdkservices.Events       `optional:"true"`
 	Integrations_ sdkservices.Integrations `optional:"true"`
+	Triggers_     sdkservices.Triggers     `optional:"true"`
 	OAuth_        sdkservices.OAuth        `optional:"true"`
 	Projects_     sdkservices.Projects     `optional:"true"`
 	Runtimes_     sdkservices.Runtimes     `optional:"true"`
 	Secrets_      sdkservices.Secrets      `optional:"true"`
 	Sessions_     sdkservices.Sessions     `optional:"true"`
 	Store_        sdkservices.Store        `optional:"true"`
-	Triggers_     sdkservices.Triggers     `optional:"true"`
 }
 
 var _ sdkservices.Services = &fxServices{}
@@ -34,10 +34,10 @@ func (s *fxServices) Dispatcher() sdkservices.Dispatcher     { return s.Dispatch
 func (s *fxServices) Envs() sdkservices.Envs                 { return s.Envs_ }
 func (s *fxServices) Events() sdkservices.Events             { return s.Events_ }
 func (s *fxServices) Integrations() sdkservices.Integrations { return s.Integrations_ }
+func (s *fxServices) Triggers() sdkservices.Triggers         { return s.Triggers_ }
 func (s *fxServices) OAuth() sdkservices.OAuth               { return s.OAuth_ }
 func (s *fxServices) Projects() sdkservices.Projects         { return s.Projects_ }
 func (s *fxServices) Runtimes() sdkservices.Runtimes         { return s.Runtimes_ }
 func (s *fxServices) Secrets() sdkservices.Secrets           { return s.Secrets_ }
 func (s *fxServices) Sessions() sdkservices.Sessions         { return s.Sessions_ }
 func (s *fxServices) Store() sdkservices.Store               { return s.Store_ }
-func (s *fxServices) Triggers() sdkservices.Triggers         { return s.Triggers_ }
