--- internal/backend/muxes/muxes.go
@@ -1,5 +1,0 @@
-package muxes
-
-import "net/http"
-
-type Muxes struct{ Auth, NoAuth *http.ServeMux }
