+++ internal/backend/svc/index.html
@@ -1,25 +1,37 @@
 <!DOCTYPE html>
 <html>
-
-<head>
+  <head>
     <title>autokitteh</title>
-    <link rel="apple-touch-icon" sizes="180x180" href="/static/apple-touch-icon.png" />
-    <link rel="icon" type="image/png" sizes="32x32" href="/static/favicon-32x32.png" />
-    <link rel="icon" type="image/png" sizes="16x16" href="/static/favicon-16x16.png" />
+    <link
+      rel="apple-touch-icon"
+      sizes="180x180"
+      href="/static/apple-touch-icon.png"
+    />
+    <link
+      rel="icon"
+      type="image/png"
+      sizes="32x32"
+      href="/static/favicon-32x32.png"
+    />
+    <link
+      rel="icon"
+      type="image/png"
+      sizes="16x16"
+      href="/static/favicon-16x16.png"
+    />
     <link rel="manifest" href="/static/site.webmanifest" />
     <link rel="stylesheet" type="text/css" href="/static/styles.css" />
     <meta name="robots" content="noindex, nofollow" />
-</head>
+  </head>
 
-<body>
+  <body>
     <ul class="navbar-container" style="list-style-type: none">
-        <li class="navbar-logo">
-            <img class="banner" src="/static/banner.svg" alt="autokitteh" />
-        </li>
+      <li class="navbar-logo">
+        <img class="banner" src="/static/banner.svg" alt="autokitteh" />
+      </li>
     </ul>
     <ul>
-        <li><a href="/i">Integrations</a></li>
+      <li><a href="/i">Integrations</a></li>
     </ul>
-</body>
-
+  </body>
 </html>
